Add DNS maps
Add basic support for mapping DNS domains to locations.

+++ lib/python2.6/aquilon/server/commands/del_dns_domain.py
@@ -48,6 +48,10 @@ class CommandDelDnsDomain(BrokerCommand):
             raise ArgumentError("DNS domain %s cannot be deleted while still "
                                 "in use." % dns_domain)
 
+        if dbdns_domain.dns_maps:
+            raise ArgumentError("{0} is still mapped to locations and cannot "
+                                "be deleted.".format(dbdns_domain))
+
         session.delete(dbdns_domain)
         session.flush()
